+++ sphinx_needs_enterprise/services/codebeamer.py
@@ -1,4 +1,5 @@
 import json
+import time
 
 from sphinx_needs_enterprise.extensions.extension import ServiceExtension
 from sphinx_needs_enterprise.util import dict_undefined_set
@@ -23,7 +24,7 @@ DEFAULT_CONTENT = """
 
 
 class CodebeamerService(ServiceExtension):
-    options = ["query", "prefix", "raw", "wiki2html"]
+    options = ["query", "prefix", "raw", "wiki2html", "wiki2html_id", "cb_request_delay_ms"]
 
     def __init__(self, app, name, config, **kwargs):
         self.app = app
@@ -37,6 +38,8 @@ class CodebeamerService(ServiceExtension):
         dict_undefined_set(config, "content", DEFAULT_CONTENT)
         dict_undefined_set(config, "raw", "False")
         dict_undefined_set(config, "wiki2html", "True")
+        dict_undefined_set(config, "wiki2html_id", 2)
+        dict_undefined_set(config, "cb_request_delay_ms", 0)
 
         mappings_default = {
             "id": ["id"],
@@ -63,6 +66,12 @@ class CodebeamerService(ServiceExtension):
         raw = options.get("raw", str(self.config["raw"]))
         options["raw"] = raw  # Just to be sure that there is a value
 
+        cb_request_delay_ms = options.get("cb_request_delay_ms", self.config["cb_request_delay_ms"])
+        options["cb_request_delay_ms"] = cb_request_delay_ms
+
+        wiki2html_id = options.get("wiki2html_id", self.config["wiki2html_id"])
+        options["wiki2html_id"] = wiki2html_id
+
         params = self._prepare_request(options)
 
         request_params = {
@@ -75,10 +84,13 @@ class CodebeamerService(ServiceExtension):
                 "descFormat": "HTML",
             },
         }
-
-        answer = self._send_request(request_params)
+        answer = self._send_request(request_params, params["cert_abspath"])
         data = answer.json()["items"]
         for datum in data:
+            delay = cb_request_delay_ms / 1000
+            if delay:
+                time.sleep(delay)
+
             # Be sure "description" is set and valid
             if "description" not in datum or datum["description"] is None:
                 datum["description"] = ""
@@ -86,7 +98,9 @@ class CodebeamerService(ServiceExtension):
                 # Transform the Codebeamer wiki syntax to HTML.
                 # Must be done by an API request for each item.
                 url = options.get("url", self.url)
-                url = url + "/api/v3/projects/2/wiki2html"
+                wiki2html_id = options.get("wiki2html_id", 2)
+
+                url = url + f"/api/v3/projects/{wiki2html_id}/wiki2html"
 
                 auth = (options.get("user", self.user), options.get("password", self.password))
 
@@ -98,7 +112,7 @@ class CodebeamerService(ServiceExtension):
                     "data": json.dumps({"markup": datum["description"]}),
                 }
 
-                wiki2html_answer = self._send_request(wiki2html_params)
+                wiki2html_answer = self._send_request(wiki2html_params, params["cert_abspath"])
                 datum["description"] = wiki2html_answer.text
 
         need_data = self._extract_data(data, options)