+++ docs/changelog.rst
@@ -1,9 +1,17 @@
 Changelog
 =========
 
+1.0.2
+-----
+**Released**: under developement
+
 1.0.1
 -----
-**Released**: under development
+**Released**: 01.02.2023
+
+* Improvement: Added :ref:`cb_ssl_cert_abspath` to allow self signed certificates for codebeamer service. This uses the verify option in the `requests library <https://requests.readthedocs.io/en/latest/api/>`.
+* Improvement: Added :ref:`cb_request_delay_ms` to configure a delay between codebeamer API :ref:`cb_wiki2html` requests.
+* Improvement: Added :ref:`cb_wiki2html_id` to change the project id used when using :ref:`cb_wiki2html` 
 
 1.0.0
 -----